refactor: waiting에서 `join`으로 라우팅 이름을 수정한다 #28
