+++ src/app/(mood)/join/components/FunnelContext.tsx
