+++ src/app/(mood)/join/components/PopupHeader.tsx
