+++ src/app/(mood)/join/funnels/step1/components/MbtiSection.tsx
