+++ src/app/(mood)/join/funnels/step1/components/WaitingHeader.tsx
