+++ src/app/(mood)/join/page.tsx
