--- src/app/(sub)/join/page.tsx
@@ -1,10 +1,0 @@
-import { JoinHeader } from '@/components/Header';
-
-export default function Join() {
-  return (
-    <>
-      <JoinHeader />
-      <h1>Join</h1>
-    </>
-  );
-}
